requirements.txt
Flask
requests
pycryptodome
protobuf
Werkzeug
gunicorn
python-dotenv